Move server health checks into workspace api module
