+++ wsmaster/che-core-api-workspace/src/test/java/org/eclipse/che/api/workspace/server/hc/TerminalHttpConnectionServerCheckerTest.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.workspace.infrastructure.docker.server;
+package org.eclipse.che.api.workspace.server.hc;
 
 import org.mockito.Mock;
 import org.mockito.testng.MockitoTestNGListener;
